refactor(mail): move setupMailAccount to MailApi and update payload structure
