+++ src/mail/api.ts
@@ -20,6 +20,7 @@ import {
   UpdateEmailRequest,
   ListEmailsQuery,
   EmailDomainsResponse,
+  SetupMailAccountPayload,
 } from './types';
 
 export class MailApi {
@@ -182,6 +183,16 @@ export class MailApi {
    */
   async getActiveDomains(): Promise<EmailDomainsResponse> {
     return this.client.get('/email/domains', this.headers());
+  }
+
+  /**
+   * Sets up a mail account for the user
+   *
+   * @param payload - Set of details for mail account setup
+   * @returns A promise that resolves with the created mail account address
+   */
+  async setupMailAccount(payload: SetupMailAccountPayload): Promise<{ address: string }> {
+    return this.client.post('/users/me/mail-account', payload, this.headers());
   }
 
   /**
